Created a simple program that have different conditions/statement
**Description**
Created a simple program that determines if the given data type are odd or even prime number or not or if its a palindrome.
"Implementation**

1.
[image](https://github.com/user-attachments/assets/a001588c-a0a7-4001-aa7d-213db1e40777)

2.
[image](https://github.com/user-attachments/assets/e5b6e79c-9fba-44fd-9965-5d9be270567)

"Conclusion"

In conclusion. The program that I created checks if a given data are accurately checking if the numbers given is odd or even, prime number or not and if it is a palindrome it breaks the data into parts and uses easy checks for its condition. Thus the program gives clear results.